+++ app/src/main/java/org/mate/exploration/genetic/fitness/CrashDistanceFitnessFunction.java
@@ -13,16 +13,25 @@ import java.util.List;
  */
 public class CrashDistanceFitnessFunction<T> implements IActionFitnessFunction<T> {
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public double getFitness(IChromosome<T> chromosome) {
         return getNormalizedFitness(chromosome);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public boolean isMaximizing() {
         return false;
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public double getNormalizedFitness(IChromosome<T> chromosome) {
 
@@ -39,7 +48,7 @@ public class CrashDistanceFitnessFunction<T> implements IActionFitnessFunction<T
      * {@inheritDoc}
      */
     @Override
-    public double getFitness(IChromosome<T> chromosome, int actions) {
+    public float getFitness(IChromosome<T> chromosome, int actions) {
         return getNormalizedFitness(chromosome, actions);
     }
 
@@ -47,12 +56,12 @@ public class CrashDistanceFitnessFunction<T> implements IActionFitnessFunction<T
      * {@inheritDoc}
      */
     @Override
-    public double getNormalizedFitness(IChromosome<T> chromosome, int actions) {
+    public float getNormalizedFitness(IChromosome<T> chromosome, int actions) {
 
-        final double crashDistance = FitnessUtils.getFitness(chromosome, actions,
+        final float crashDistance = FitnessUtils.getFitness(chromosome, actions,
                 FitnessFunction.CRASH_DISTANCE);
 
-        if (crashDistance == 0.0d) { // We can terminate the search once we have reproduced the crash.
+        if (crashDistance == 0.0f) { // We can terminate the search once we have reproduced the crash.
             ConditionalTerminationCondition.satisfiedCondition();
         }
 
@@ -63,11 +72,11 @@ public class CrashDistanceFitnessFunction<T> implements IActionFitnessFunction<T
      * {@inheritDoc}
      */
     @Override
-    public List<Double> getNormalizedFitnessVector(IChromosome<T> chromosome) {
+    public List<Float> getNormalizedFitnessVector(IChromosome<T> chromosome) {
 
-        final List<Double> crashDistances = FitnessUtils.getCrashDistanceVector(chromosome);
+        final List<Float> crashDistances = FitnessUtils.getCrashDistanceVector(chromosome);
 
-        if (crashDistances.contains(0.0d)) {
+        if (crashDistances.contains(0.0f)) {
             // We can terminate the search once we have reproduced the crash.
             ConditionalTerminationCondition.satisfiedCondition();
         }
